# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "MySQL Speaker",
    "description": "API Documentation of MySQL Speaker"
  },
  "servers": [
    {
      "url": "https://mysqlspeaker.online/",
      "description": ""
    }
  ],
  "paths": {
    "/api/user": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for user creation.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "invitationCode": {
                    "example": "any"
                  },
                  "newGroupName": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/auth": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for user authentication.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for verifying user authentication.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/user/guestAuth": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for guest user authentication.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "account": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "dbHost": {
                    "example": "any"
                  },
                  "userName": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data": {
      "put": {
        "tags": [
          "Update"
        ],
        "description": "Endpoint to update individual data.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "data": {
                    "example": "any"
                  },
                  "where": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Update"
        ],
        "description": "Endpoint to insert new row data.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "values": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Delete"
        ],
        "description": "Endpoint for deleting specific data.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "where": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/table/column": {
      "put": {
        "tags": [
          "Update"
        ],
        "description": "Endpoint to add a new column to a table.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "columnName": {
                    "example": "any"
                  },
                  "columnType": {
                    "example": "any"
                  },
                  "columnOption": {
                    "example": "any"
                  },
                  "defaultValue": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Update"
        ],
        "description": "Endpoint to delete a specified column from a table.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized access, no token provided or invalid token"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dbName": {
                    "example": "any"
                  },
                  "table": {
                    "example": "any"
                  },
                  "columnName": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
      }
    },
...
  }
}